Dev Commit. 	1.rewrite @func DAGCircuit::add_instructionNode_end() 	2.finish sabre_routing v1

+++ include/sabre_core.h
@@ -2,5 +2,5 @@
 
 using qubit_t = int;
 using cbit_t = int;
-using node_pos_t = int;
+using node_pos_t = unsigned long;
 using edge_pos_t = std::pair<node_pos_t, node_pos_t>;
